modified:   i3wm/i3/config 	modified:   i3wm/i3/scripts/script.sh

+++ i3wm/i3/scripts/script.sh
@@ -1,12 +1,23 @@
 #!/bin/bash
 
+detail_vol(){
+    CheckIdVol=$(wpctl status | grep "Built-in Audio Analog Stereo" | awk '{print $3}'| head -n1 |sed 's/\.//')
+    VOL=$(wpctl get-volume $CheckIdVol | awk '{print ($3 ? "MUTE" : int($2*100))}')
+    echo $VOL
+}
 volume_up() {
-    pactl set-sink-volume @DEFAULT_SINK@ +5%
-    pkill -SIGRTMIN+10 i3blocks
+    if [ $(detail_vol) -ge 100 ];then
+        dunstify -t 1000 -h string:x-dunst-stack-tag:volume -h int:value:100 "Progress" "100%"
+    else 
+        pactl set-sink-volume @DEFAULT_SINK@ +5%
+        pkill -SIGRTMIN+10 i3blocks
+        dunstify -t 1000 -h string:x-dunst-stack-tag:volume -h int:value:$(detail_vol) "Progress" "$(detail_vol)%"
+    fi
 }
 volume_down() {
     pactl set-sink-volume @DEFAULT_SINK@ -5%
     pkill -SIGRTMIN+10 i3blocks
+    dunstify -t 1000 -h string:x-dunst-stack-tag:volume -h int:value:$(detail_vol) "Progress" "$(detail_vol)%"
 }
 logout(){
     options="Logout\nSuspend\nReboot\nShutdown\nLock"
@@ -35,7 +46,7 @@ logout(){
 switch-workspaces(){
     workspaces=$(i3-msg -t get_workspaces | jq -r '.[] | select(.focused == false) | .name')
     if [ -z "$workspaces" ]; then
-        notify-send -t 5000 "Tidak ada workspace yang aktif"
+        dunstify -t 2000 -h string:x-dunst-stack-tag:workspace "Tidak ada workspace yang non-aktif"
         exit 0
     fi
     selected=$(echo "$workspaces" | rofi -dmenu -p "Workspace")
